Predators only copy if their hunger is 0
The simulation seems to work a bit better now, I only let the wolves
copy if their hunger is 0.

What ends up happening is:

- most of the wolves die off pretty quickly
- then lots of bunnies live
- then those lone survivor wolves gobble them up (they’re pretty
densely packed).

It kind of gets to a steady state here, not a lot of wolves and not a
lot of bunnies (but more bunnies than wolves).

I’m noticing that it’s really hard to tell what’s going on, in a way.

- I don’t know how many wolves or bunnies are alive/dead
- I don’t know how it changes over time
- I want to start tweaking all kinds of rates, but I don’t know how
they’ll affect things, or how I’ll compare them

This is probably where math wins out a little bit. They actual math for
predator prey reveals values for those things. My simulation is
probably an approximation (at best) of that formula, but it’s not an
official implementation of it.